Add null check when deleting a sequence. (#49)
* Add null check when deleting a sequence.

* Added a unit test for good measure

+++ remote-falcon-api/src/main/java/com/remotefalcon/api/service/ControlPanelService.java
@@ -451,7 +451,9 @@ public class ControlPanelService {
   public ResponseEntity<?> deleteSequence(Long sequenceKey) {
     TokenDTO tokenDTO = this.authUtil.getJwtPayload();
     Playlist playlist = this.playlistRepository.findByRemoteTokenAndSequenceKey(tokenDTO.getRemoteToken(), sequenceKey);
-    this.playlistRepository.delete(playlist);
+    if(playlist != null) {
+      this.playlistRepository.delete(playlist);
+    }
     return ResponseEntity.status(200).build();
   }
